requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
bidict==0.23.1
click==8.1.8
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-decouple==3.8
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
wsproto==1.2.0
